update
all the code from dealer move to game script file.
