+++ src/components/Game/game.css
@@ -20,3 +20,15 @@
   align-items: center;
   text-align: center;
 }
+
+.dealer-game-content {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  align-content: center;
+  justify-content: flex-start;
+  align-items: center;
+  margin-top: 20px;
+  align-items: center;
+  text-align: center;
+}
