--- src/components/Game/menu.css
@@ -1,8 +1,0 @@
-.drawn-cards {
-  display: inline-flex;
-  justify-content: center;
-}
-
-.divider {
-  margin-bottom: 25px;
-}
