--- src/hooks/usaGame.jsx
@@ -1,1 +1,0 @@
-export default function useGame() {}
